+++ src/main/java/com/epam/web/logic/service/account/AccountService.java
@@ -3,14 +3,15 @@ package com.epam.web.logic.service.account;
 import com.epam.web.exception.ServiceException;
 import com.epam.web.model.entity.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
     List<Account> getAll() throws ServiceException;
 
-    int getBalance(long id) throws ServiceException;
+    float getBalance(long id) throws ServiceException;
 
-    void addMoneyById(int money, long id) throws ServiceException;
+    void addMoneyById(float money, long id) throws ServiceException;
 
     void unblockById(long id) throws ServiceException;
